requirements.txt
colorz==1.0.3
Flask==1.1.2
numpy==1.18.4
Pillow==2.2.1
svgwrite==1.4
urllib3==1.25.9
